Modifying board to include jumpers to change pin 1 function
Following a recent video on Adrians Digital Basement where this module was used in a machine that required Motorola bus mode to function, and seeing the modification made with a bodge wire, I thought it would be good to include some solder bridges that could be used to implement a cleaner solution to make this module more universal.

Two 3-way solder jumpers have been included that allow pin 1 of the RTC chip to be routed either directly to pin 1 of the socket/header, or to a second 3-way solder jumper which allows selecting between a hard pull up or down. Hard pull up/down is used by recommendation of the datasheet for the MC146818A and bq3285.